refactor: extract dependent resources code to separate file
